pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "streamsniper"
version = "2.0.0"
description = "YouTube & video downloader with a dark-themed GUI"
license = "GPL-3.0-only"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.01.01",
    "Pillow>=10.0",
]

[tool.setuptools.packages.find]
include = ["streamsniper*"]
